fix(core): aggregated the various internal reflection functions under ValueInspector
ValueInspector can be imported from core/lib/io and used to inspect the type and characteristics of
objects and primitive types. It's mainly used internally by Serenity/JS modules.

Related tickets: re #2223
