+++ packages/core/src/io/reflection/index.ts
@@ -1,3 +1,2 @@
 export * from './has';
-export * from './isPrimitive';
-export * from './typeOf';
+export * from './ValueInspector';
