+++ packages/core/src/io/stringified.ts
@@ -2,7 +2,7 @@ import { inspect } from 'util';
 
 import type { Answerable } from '../screenplay/Answerable';
 import { Question } from '../screenplay/Question';
-import { isPlainObject } from './isPlainObject';
+import { ValueInspector } from './reflection/ValueInspector';
 
 const indentationPrefix = '  ';
 
@@ -33,7 +33,7 @@ export function stringified(value: Answerable<any>, config?: StringifyConfig): s
         return stringifiedArray(value, { indentationLevel, inline, markQuestions });
     }
 
-    if (isAPromise(value)) {
+    if (ValueInspector.isPromise(value)) {
         return markAs('Promise', markQuestions);
     }
 
@@ -41,25 +41,25 @@ export function stringified(value: Answerable<any>, config?: StringifyConfig): s
         return markAs(value.toString(), markQuestions);
     }
 
-    if (isADate(value)) {
+    if (ValueInspector.isDate(value)) {
         return value.toISOString();
     }
 
-    if (hasItsOwnToString(value)) {
+    if (ValueInspector.hasItsOwnToString(value)) {
         return value.toString();
     }
 
-    if (isInspectable(value)) {
+    if (ValueInspector.isInspectable(value)) {
         return value.inspect();
     }
 
-    if (isAFunction(value)) {
+    if (ValueInspector.isFunction(value)) {
         return hasName(value)
             ? value.name
             : markAs(`Function`, true);
     }
 
-    if (! hasCustomInspectionFunction(value) && isPlainObject(value) && isSerialisableAsJSON(value)) {
+    if (! ValueInspector.hasCustomInspectionFunction(value) && ValueInspector.isPlainObject(value) && isSerialisableAsJSON(value)) {
         return stringifiedToJson(value, { indentationLevel, inline, markQuestions });
     }
 
@@ -122,70 +122,12 @@ function isDefined(v: Answerable<any>) {
 }
 
 /**
- * Checks if the value defines its own `toString` method
- *
- * @param v
- */
-function hasItsOwnToString(v: Answerable<any>): v is { toString: () => string } {
-    return typeof v === 'object'
-        && !! (v as any).toString
-        && typeof (v as any).toString === 'function'
-        && ! isNative((v as any).toString);
-}
-
-/**
- * Checks if the value defines its own `inspect` method
- *
- * @param v
- */
-function isInspectable(v: Answerable<any>): v is { inspect: () => string } {
-    return !! (v as any).inspect && typeof (v as any).inspect === 'function';
-}
-
-/**
- * Checks if the value is a {@apilink Date}
- *
- * @param v
- */
-function isADate(v: Answerable<any>): v is Date {
-    return v instanceof Date;
-}
-
-/**
- * Checks if the value is a {@apilink Promise}
- *
- * @param v
- */
-function isAPromise<T>(v: Answerable<T>): v is Promise<T> {
-    return typeof v === 'object'
-        && 'then' in v;
-}
-
-/**
- * Checks if the value is a named {@apilink Function}
- *
- * @param v
- */
-function isAFunction(v: any): v is Function {       // eslint-disable-line @typescript-eslint/ban-types
-    return Object.prototype.toString.call(v) === '[object Function]';
-}
-
-/**
  * Checks if the value is has a property called 'name' with a non-empty value.
  *
  * @param v
  */
 function hasName(v: any): v is { name: string } {
     return typeof (v as any).name === 'string' && (v as any).name !== '';
-}
-
-/**
- * Checks if the value defines its own [`inspect` method](https://nodejs.org/api/util.html#util_util_inspect_custom)
- *
- * @param v
- */
-function hasCustomInspectionFunction(v: Answerable<any>): v is object { // eslint-disable-line @typescript-eslint/ban-types
-    return v && v[Symbol.for('nodejs.util.inspect.custom')];
 }
 
 /**
@@ -201,42 +143,4 @@ function isSerialisableAsJSON(v: any): v is object {    // eslint-disable-line @
     } catch {
         return false;
     }
-}
-
-/**
- * Inspired by https://davidwalsh.name/detect-native-function
- *
- * @param v
- */
-function isNative(v: any): v is Function {  // eslint-disable-line @typescript-eslint/ban-types
-
-    const
-        toString        = Object.prototype.toString,    // Used to resolve the internal `{@apilink Class}` of values
-        fnToString      = Function.prototype.toString,  // Used to resolve the decompiled source of functions
-        hostConstructor = /^\[object .+?Constructor]$/; // Used to detect host constructors (Safari > 4; really typed array specific)
-
-    // Compile a regexp using a common native method as a template.
-    // We chose `Object#toString` because there's a good chance it is not being mucked with.
-    const nativeFunctionTemplate = new RegExp(
-        '^' +
-        // Coerce `Object#toString` to a string
-        String(toString)
-        // Escape any special regexp characters
-            .replaceAll(/[$()*+./?[\\\]^{|}]/g, '\\$&')
-            // Replace mentions of `toString` with `.*?` to keep the template generic.
-            // Replace thing like `for ...` to support environments like Rhino which add extra info
-            // such as method arity.
-            .replaceAll(/toString|(function).*?(?=\\\()| for .+?(?=\\])/g, '$1.*?') +
-        '$',
-    );
-
-    const type = typeof v;
-    return type === 'function'
-        // Use `Function#toString` to bypass the value's own `toString` method
-        // and avoid being faked out.
-        ? nativeFunctionTemplate.test(fnToString.call(v))
-        // Fallback to a host object check because some environments will represent
-        // things like typed arrays as DOM methods which may not conform to the
-        // normal native pattern.
-        : (v && type === 'object' && hostConstructor.test(toString.call(v))) || false;
 }
